pyproject.toml
[project]
name = "garage-storage"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.0",
    "flask-login>=0.6.3",
    "flask-migrate>=4.0.7",
    "flask-sqlalchemy>=3.1.1",
    "flask-uploads>=0.2.1",
    "flask-wtf>=1.2.2",
    "google-genai>=0.3.0",
    "google-generativeai>=0.8.3",
    "jinja2>=3.1.5",
    "loguru>=0.7.3",
    "openai>=1.58.1",
    "pandas>=2.2.3",
    "pillow>=11.0.0",
    "pyopenssl>=24.3.0",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "qrcode>=8.0",
    "requests>=2.32.3",
    "rich>=13.9.4",
    "sqlmodel>=0.0.22",
    "werkzeug>=3.1.3",
]

[dependency-groups]
dev = [
    "ruff>=0.8.4",
]
